+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## 1.90
 
+- Improve Network Registry smart contract to allow 1-to-many node registration, add enable/disable make targets ([#4008](https://github.com/hoprnet/hoprnet/pull/4091))
+
 ---
 
 <a name="1.89"></a>